Fixes SlipperyComponent mispredict
