+++ Content.Client/GameObjects/Components/Movement/SlipperyComponent.cs
@@ -7,17 +7,5 @@ namespace Content.Client.GameObjects.Components.Movement
     [ComponentReference(typeof(SharedSlipperyComponent))]
     public class SlipperyComponent : SharedSlipperyComponent
     {
-        public override void HandleComponentState(ComponentState? curState, ComponentState? nextState)
-        {
-            base.HandleComponentState(curState, nextState);
-
-            if (curState is not SlipperyComponentState state) return;
-
-            _slippery = state.Slippery;
-            _intersectPercentage = state.IntersectPercentage;
-            _paralyzeTime = state.ParalyzeTime;
-            _requiredSlipSpeed = state.RequiredSlipSpeed;
-            _launchForwardsMultiplier = state.LaunchForwardsMultiplier;
-        }
     }
 }
