+++ Content.Server/GameObjects/Components/Movement/SlipperyComponent.cs
@@ -2,9 +2,6 @@
 using Content.Shared.GameObjects.Components.Movement;
 using Robust.Server.GameObjects;
 using Robust.Shared.GameObjects;
-using Robust.Shared.Players;
-using Robust.Shared.Serialization.Manager.Attributes;
-using Robust.Shared.ViewVariables;
 
 namespace Content.Server.GameObjects.Components.Movement
 {
@@ -12,13 +9,6 @@ namespace Content.Server.GameObjects.Components.Movement
     [ComponentReference(typeof(SharedSlipperyComponent))]
     public class SlipperyComponent : SharedSlipperyComponent
     {
-        /// <summary>
-        ///     Path to the sound to be played when a mob slips.
-        /// </summary>
-        [ViewVariables]
-        [DataField("slipSound")]
-        public string SlipSound { get; set; } = "/Audio/Effects/slip.ogg";
-
         protected override void OnSlip()
         {
             if (!string.IsNullOrEmpty(SlipSound))
@@ -26,11 +16,6 @@ namespace Content.Server.GameObjects.Components.Movement
                 EntitySystem.Get<AudioSystem>()
                     .PlayFromEntity(SlipSound, Owner, AudioHelpers.WithVariation(0.2f));
             }
-        }
-
-        public override ComponentState GetComponentState(ICommonSession player)
-        {
-            return new SlipperyComponentState(ParalyzeTime, IntersectPercentage, RequiredSlipSpeed, LaunchForwardsMultiplier, Slippery);
         }
     }
 }
